add zaaack.markdown-editor and dan-c-underwood.arm
1.zaaack.markdown-editor
Markdown Editor — A full-featured WYSIWYG editor for markdown

2.dan-c-underwood.arm
ARM assembly highlighting for Visual Studio Code
Adds syntax highlighting for the ARM Assembly language to Visual Studio Code. This extension was originally written to support the ARM instruction set used in teaching at the University of Manchester so availability of instructions on platforms is not guaranteed.

The extension is a converted version of the syntax highlighting package for the Atom editor.